fix : storeId int -> Long

+++ src/main/java/com/example/delivery/repository/store/StoreRepository.java
@@ -7,11 +7,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface StoreRepository extends JpaRepository<StoreEntity, Integer> {
+public interface StoreRepository extends JpaRepository<StoreEntity, Long> {
     // 유저에 해당하는 가게 목록 조회
     List<StoreEntity> findByUser(UserEntity user);
     // 가게 ID로 찾기
-    Optional<StoreEntity> findById(int storeId);
+    Optional<StoreEntity> findById(Long storeId);
     //폐업안된 가게 찾기
     List<StoreEntity> findByClosedFalse();
 }
